backend/scripts/requirements.txt
# Python Dependencies for QR System
# Install with: pip install -r requirements.txt

# QR Code Generation
qrcode[pil]>=7.4.2

# Image Processing
Pillow>=10.0.0

# PowerPoint/PPTX Processing (for SDS templates)
python-pptx>=0.6.21
